Fixed a problem: - When registering and then quit the app, we are saved in the database but we have no section which is problematic
